requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
cron-descriptor==1.4.3
Django==4.1.13
django-auth-email==0.0.1a5
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-ipware==5.0.1
django-rest-authemail==2.1.7
django-timezone-field==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
exception==0.1.0
flower==2.0.1
humanize==4.9.0
idna==3.4
kombu==5.3.5
numpy==1.26.4
Pillow==10.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pymongo==4.6.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
tornado==6.4
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
